mcpdemo/pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.28",
    "mcp-use>=1.3.7",
    "python-dotenv>=1.0.0",
]
